# OrderlyBound: specialized BoundRemainingBoards for orderly trees