--- a/boggle/split_order.py
+++ b/boggle/split_order.py
@@ -1,5 +1,7 @@
 """Order in which to split cells. Middle then edges then corners."""
 
+from typing import Sequence
+
 SPLIT_ORDER_33 = (4, 5, 3, 1, 7, 0, 2, 6, 8)
 
 
@@ -50,7 +52,7 @@ SPLIT_ORDER_44 = tuple(
 assert len(SPLIT_ORDER_44) == 16
 assert len(set(SPLIT_ORDER_44)) == 16
 
-SPLIT_ORDER = {
+SPLIT_ORDER: dict[tuple[int, int], Sequence[int]] = {
     (2, 2): (0, 1, 2, 3),
     (2, 3): (0, 1, 2, 3, 4, 5),
     (3, 3): SPLIT_ORDER_33,