I got this idea from the visualizations in my [recent blog post][1]:

![image](https://github.com/user-attachments/assets/83174406-bf4d-44f8-b293-1a8944dfa773)

- Start with the leftmost blue and green choice cells.
- Sum their bounds. If it's less than the cutoff, you're done.
- If not, try each possible blue choice and continue down that path to all the next choices.

You maintain a stack of nodes for choices on each cell and steadily remove early choices and replace them with more later choices. You're able to prune as you go, just like with BoundRemainingBoards. It's kind of like a BFS that takes advantage of the "orderly" structure of the tree. It should only visit each node once, producing the max/no-mark bound for each board if you get all the way to a leaf. If you have an orderly tree, it's a drop-in replacement for BoundRemainingBoards that's ~10x faster. If not, it doesn't work (it would require backtracking).

This is a huge speedup for BoundRemainingBoards. I'd hoped that it would be a big enough speedup that I could ditch lifting entirely. That doesn't quite seem to be the case. Going straight from `build_tree` to `orderly_bound` gets me the fastest time I've seen yet on the 50 board test set, 21s vs 23s before. And it uses less RAM. But it's slower than lifting for the hardest 3x4 board classes by around 30%.

207836 (hard):

- old switch=0: 12.94 = 1.16 + 11.46 (706MB)
- old switch=2: 5.29 = 0.96 + 0.36 + 3.6 (806MB)
- old switch=4: 3.30 = 0.95 + 0.67 + 0.38 + 0.68 + 0.30 (1.41GB)
- orderly bound: 3.86 = 1.01 + 2.64 (707MB)
- orderly switch=2: 2.62 = 1 + 0.39 + 1.22

234556 (best):

- brb s=0: 310s
- brb s=4: 23.66s = 2.4s + 0.98s + 0.90s + 1.99s + 15.4s + 0.4s (2.8GB)
- new: 37.61s = 2.2s + 34.6s + 0.4s (1.2GB)
- new s=2: 20.72s = 2.2 + 0.9 + 17.2 + 0.4s (1.48GB)
- new s=3: 14.45s = 2.2 + 0.9 + 0.9 + 10 + 0.4s (1.76GB)
- new s=4: 11.34s = 2.1 + 0.9 + 0.9 + 2 + 5 + 0.4s (2.8GB)

524960 (hardest):

- brb s=4: 46.66s = 2.1s + 0.6s + 0.6s + 1.4s + 38s + 3.4s (2.17G)
- brb s=5: 27.12s = 1.9s + 0.6s + 0.6s + 1.5s + 4s + 13.5s + 3.4s (3.81G)
- new: 53.8s = 1.9s + 48.5s + 3.4s (1.13GB)
- new s=2: 36.66s = 1.9 + 0.74 + 30.62 + 3.4s (1.29GB)
- new s=3: 25.47s = 1.8 + 0.6 + 0.6 + 19 + 3.4s (1.61GB)
- new s=4: 18.53s = 2 + 0.6 + 0.6 + 1.3 + 10.6 + 3.4 (2.18G)
- new s=5: 17.8 = 1.8 + 0.6 + 0.6 + 1.2 + 4.3 + 5.8 + 3.5 (3.84G)

So the fastest approach winds up being lifting + orderly bounds, at least for hard boards. A lifted orderly tree is still orderly. It just means that I wouldn't be able to refactor EvalNodes without re-implementing lifting. This shifts the curve for switchover levels in a good way: switchover=5 is only marginally faster for 524960 whereas previously it was a clear win. Or alternatively you can lower the switchover level to get a similar runtime but use a lot less memory.

To-do before merging:

- [x] Get it working with lifted trees or pull it out of breaker.
- [x] Figure out why ordery_bound crashes after 4 lifts

[1]: https://www.danvk.org/2025/02/21/orderly-boggle.html